- Moved feature importance's code to its own file (at this moment   feature_importance.py) Waiting for a better name...
